feat: enhance cache settings UI and functionality
- Updated cache settings UI for improved readability and usability.
- Increased font sizes and adjusted spacing for headings and descriptions.
- Refined radio button implementation for cache policy selection with better visual feedback.
- Added a new label for saved cache policy confirmation.
- Improved cache policy handling logic in the JavaScript component.
- Removed unused per-key TTL summary section from the UI.

feat: implement CLI API key management for admins

- Created AdminCliKeyController for managing CLI API keys across projects.
- Added endpoints for listing all API keys, issuing new keys, and toggling key status.
- Introduced DTOs for API key requests and responses to streamline data handling.
- Implemented a custom authentication failure handler for better user experience on login failures.
- Enhanced the CLI key management UI with a slideout for creating new keys and improved error handling.

+++ src/main/java/com/salkcoding/oswl/auth/entity/CacheSetting.java
@@ -19,7 +19,7 @@ public class CacheSetting {
     private String cacheKey;
 
     @Column(name = "ttl_seconds", nullable = false)
-    private int ttlSeconds;
+    private long ttlSeconds;
 
     @Column(name = "last_cleared_at")
     private LocalDateTime lastClearedAt;
